One more syntax rule in search-ahead from quote inside quote-enclosed tag.
We look ahead to determine is such a quote sign should be taken as the
end of the parameter.
The rule was: yes, if it is followed by either space and quote or
[space and] tag-end symbol.
Now there is one more option: yes, if followed by anything but
tag-begin and quote until first tag-end.