Initial (raw) version of buildGUI method in TagParser, and used for one tag dialog.
This first version just uses the buildGUI method for the tag
‘paragraph-space’.
Format parameter names now also appear as titles in GUI, so for this
tag they are shortened from e.g. ‘space before paragraph’ to ‘before’,
which is OK in the context.
- buildGUI picks the syntax selected by the parser as the “current
syntax” and there is no code yet to present multiple syntax options to
the user.
- we pick up possible error from the parser but don’t show it in the
tag dialog yet.
- the decorated number widget should not show decorations combo-box if
it is limited to ‘%’ (mandatory or as option).