Handling update list of options in parameter in syntax of parser
So a new method TagParser.updateOptions.
Necessary in order to use generic tag dialog for “include” tag which
has a parameter of type “option” which is the list of extensions - a
list that can change.
Generic tag dialog for the tags: face, orientation, blend and include.
Changed type of parameter in parser for ‘include’ from string to option.