Also use parameter default value to decide if non-mandatory value should be inserted.
The get method builds a string for the tag for insertion into the
source code. For parameters of type float it now skips non-mandatory
values if either blank or equal to default value. So this is a second
use of the default value. The first was to be the default value in the
tag dialog.
Generic tag dialog extended to handle the ‘new’ tag.