Actions on formal parameters for buttons in the tag dialog.
New abstract TagAction extend Action (swing) and this first case of the
FileChooser. Formal parameters get array of TagActions and an offset
for where to action in the list of GUI fields.
FileChooser TagAction added to parser for the ‘insert’ tag.
ParameterType class for the GUI fields (rename?) also gets array of
TagActions, the array of fields itself and an offset for access into
this array.
The FileChooser needs access to the fields both when setting directory
on the dialog and for setting the appropriate field with the chosen
file. This cannot be set up when parsing (too early) so this happens
when the tag dialog is built.