Improve glyph dialog.
Matching of short font encoding id with list of encoding titles moved
to the object with font encoding / code page functions.
Include “OEM” for special handling.
When refreshing layout of tag dialog from a change of a combo-box in
the tag dialog, keep the latest forced syntax choice. Otherwise it can
“suddenly” jump to another syntax, if the parser finds a matter match
for your temporary entries.